% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_parents_from_children.R
\name{aggregate_from_children_to_parents}
\alias{aggregate_from_children_to_parents}
\title{Aggregate from children to parents using a hierarchy}
\usage{
aggregate_from_children_to_parents(
  DT,
  hierarchy,
  varnames_to_aggregate,
  varnames_to_aggregate_by,
  stop_level = 0L,
  hierarchy_id = "location_id"
)
}
\arguments{
\item{DT}{[data.table] e.g. some data table with hierarchy_id as a column}

\item{hierarchy}{[data.table] e.g. a location hierarchy with required columns: `hierarchy_id`, path_to_top_parent, level}

\item{varnames_to_aggregate}{[chr] e.g. c("cases", "deaths")}

\item{varnames_to_aggregate_by}{[chr] e.g c("year", "age_group")}

\item{stop_level}{[x] (default 0L) According to hierarchy levels, what is the top level?}

\item{hierarchy_id}{[chr] e.g. "location_id" (currently only supported option)}
}
\value{
[data.table] aggregated data.table
}
\description{
Aggregate from leaf nodes up through an (assumed MECE) hierarchy to the top
level.
}
\details{
Relies on the `children_of_parents()` function to find children of a parent
hierarchy_id e.g. location_id, then aggregates the selected columns for all
children of one parent.
}
