% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_tribble.R
\name{align_commas_in_tribble}
\alias{align_commas_in_tribble}
\title{Align a tribble defined in an R file by all the commas}
\usage{
align_commas_in_tribble(path, make_backup)
}
\arguments{
\item{path}{[chr] path to R file containing a single defined
`tibble::tribble()`}

\item{make_backup}{[lgl] whether to make a backup of the original file before
styling aligning commas}
}
\value{
[invisible character vector] the lines of the file after alignment
}
\description{
Visually creates fixed-width text columns
}
