% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{apply_comma_string_to_list}
\alias{apply_comma_string_to_list}
\title{Convert atomic list elements to comma strings}
\usage{
apply_comma_string_to_list(lst)
}
\arguments{
\item{lst}{[list] a list of (potentially) atomic vectors}
}
\value{
[list] a list with atomic elements converted to comma-separated strings
}
\description{
Facilitates arg-parsing at the command line.
}
\examples{
apply_comma_string_to_list(list(1:5, c("a")))
#> [[1]]
#> [1] "1,2,3,4,5"

#> [[2]]
#> [1] "a"
}
