% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_regional_scalars.R
\name{apply_regional_scalars}
\alias{apply_regional_scalars}
\title{Multiply data.table columns by GBD regional scalars}
\usage{
apply_regional_scalars(DT, vars_to_multiply, release_id, location_set_id)
}
\arguments{
\item{DT}{[data.table] with columns `location_id`, `year_id`, and variables
to be aggregated}

\item{vars_to_multiply}{[chr] vector of variable names in `DT` to be
multiplied by regional scalars}

\item{release_id}{[int] release_id for regional scalars}

\item{location_set_id}{[int] location_set_id for regional scalars}
}
\value{
[data.table] input `DT` with `vars_to_multiply` multiplied by
regional scalars where matched by `location_id` and `year_id`
}
\description{
Support function for aggregate_parents_from_children
}
