% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_valid_choices.R
\name{assert_all_in}
\alias{assert_all_in}
\title{Assert all elements of x are in required}
\usage{
assert_all_in(x, required)
}
\arguments{
\item{x}{[vector] some vector}

\item{required}{[vector] some vector}
}
\value{
[none] stop if any elements of x are not in required
}
\description{
Assert all elements of x are in required
}
