% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_helpers.R
\name{assert_git_diff}
\alias{assert_git_diff}
\title{Stop progress if user has uncommitted changes}
\usage{
assert_git_diff(git_uncommitted)
}
\arguments{
\item{git_uncommitted}{[character] output from git_diff function}
}
\value{
[NULL]
}
\description{
Stop progress if user has uncommitted changes
}
