% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_helpers.R
\name{assert_git_hash}
\alias{assert_git_hash}
\title{Assert that daughter scripts share a common code state with launch script}
\usage{
assert_git_hash(launch_hash, script_hash)
}
\arguments{
\item{script_hash}{[character] daughter script hash to check against main}

\item{metadata}{[list] metadata written in launch script}
}
\value{
passing statement or error message
}
\description{
Assert that daughter scripts share a common code state with launch script
}
