% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_list_elements_and_types}
\alias{assert_list_elements_and_types}
\title{Assert that a list contains all the element from a _named_ check list.}
\usage{
assert_list_elements_and_types(
  check_list,
  truth_list,
  allow_data_frame = FALSE
)
}
\arguments{
\item{check_list}{[list] named list of items to check for presence and data
type.}

\item{truth_list}{[list] named list of pre-determined items with data types.
If you don't wish to check data type, save as `NA` e.g. `list(arg1 = NA)`}

\item{allow_data_frame}{[lgl] TRUE if `check_list` is a data.frame (allows
checking column presence and type).}
}
\value{
none
}
\description{
Each `truth_list` list element name is an element that should exist in
`check_list.` The value of each `truth_list` list element is the data type
that the corresponding `check_list` element should be. If you don't wish to
check data type, save as `NA` e.g. `list(arg1 = NA)`.
}
\examples{
assert_list_element_and_type(list(a = 1, b = "2"), list(a = "double", b = "character"))
assert_list_element_and_type(mtcars, list(mpg = "double", cyl = "double"), allow_data_frame = TRUE)
}
