% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_named_list}
\alias{assert_named_list}
\title{Assert that an object is a list, and that all top level items have names}
\usage{
assert_named_list(x, allow_data_frame = FALSE)
}
\arguments{
\item{x}{[list] object to be checked}
}
\description{
Assert that an object is a list, and that all top level items have names
}
\examples{
assert_named_list(list(a = 1, b = 2))
}
