% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_no_na.R
\name{assert_no_na}
\alias{assert_no_na}
\title{Assert no NA values in select columns}
\usage{
assert_no_na(dt, varnames, verbose = FALSE)
}
\arguments{
\item{dt}{[data.table]}

\item{varnames}{[chr] vector of variable names to check for NA values}

\item{verbose}{[lgl] print success message?}
}
\value{
[none] stop if NA values found in `varnames`
}
\description{
Assert no NA values in select columns
}
