% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_scalar.R
\name{assert_scalar}
\alias{assert_scalar}
\title{Ensure an element is atomic and length 1}
\usage{
assert_scalar(x)
}
\arguments{
\item{x}{[any] some element}
}
\value{
[std_err] stop if not length 1
}
\description{
Ensure an element is atomic and length 1
}
