% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_square.R
\name{assert_square}
\alias{assert_square}
\title{Assert a data.table is square by id_varnames}
\usage{
assert_square(
  dt,
  id_varnames,
  no_na_varnames = NULL,
  verbose = FALSE,
  hard_stop = TRUE,
  stop_if_empty = TRUE
)
}
\arguments{
\item{dt}{[data.table] table of data}

\item{id_varnames}{[chr] vector of id variable names that, in combination (e.g. expand.grid) will uniquely ID all rows of data}

\item{no_na_varnames}{[chr: default NULL] optional vector of variable names that must not be NA to be considered 'square'}

\item{verbose}{[lgl: default FALSE] print success message?}

\item{hard_stop}{[lgl: default TRUE] - if non-square, stop if TRUE warn if FALSE}

\item{stop_if_empty}{[lgl: default TRUE] - if `dt` has no rows, stop if TRUE warn if FALSE}
}
\value{
[list] 2 data.tables - duplicated rows and missing rows
}
\description{
Checks:
- duplicated rows by id_var combination
- missing rows by id_var combination
}
\examples{
# fail
.chk_list <- assert_square(dt = data.table::data.table(num = 1:3, let = letters[1:3]), id_varnames = c("num", "let"), verbose = TRUE, hard_stop = FALSE)
# pass
.chk_list <- assert_square(dt = data.table::CJ(num = 1:3, let = letters[1:3]), id_varnames = c("num", "let"), verbose = TRUE)
}
