% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_x_in_y.R
\name{assert_x_in_y}
\alias{assert_x_in_y}
\title{Assert all elements of x are in y}
\usage{
assert_x_in_y(x, y)
}
\arguments{
\item{x}{[vector] some vector}

\item{y}{[vector] some vector}
}
\value{
[none] stop if any elements of x are not in y
}
\description{
Assert all elements of x are in y
}
