% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_ids.R
\name{attach_ihme_loc_id}
\alias{attach_ihme_loc_id}
\title{Use `location_id` to merge on `ihme_loc_id` from a location table.}
\usage{
attach_ihme_loc_id(
  df,
  hierarchy,
  loc_id_varname = "location_id",
  ihme_loc_varname = "ihme_loc_id"
)
}
\arguments{
\item{df}{[data.frame] of data with `location_id` column}

\item{hierarchy}{[data.frame] location table with `ihme_loc_id` and
`location_id` columns}

\item{loc_id_varname}{[chr: default "location_id"] variable name in `df` for
`location_id`}

\item{ihme_loc_varname}{[chr: default "ihme_loc_id"] variable name in `df`
for `ihme_loc_id`}
}
\value{
[data.frame] input `df` with `ihme_loc_id` column added if not
  already present
}
\description{
User control of variable names for `location_id` and `ihme_loc_id` in case
they differ from the defaults.
}
\details{
Designed to gracefully move on if `ihme_loc_id` column is already present
(since this is a standard column name).
}
