% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_ids.R
\name{attach_national_ihme_loc_id}
\alias{attach_national_ihme_loc_id}
\title{Attach `national_ihme_loc_id` column to a data.frame using `location_id`}
\usage{
attach_national_ihme_loc_id(
  df,
  hierarchy,
  allow_self_as_parent = TRUE,
  new_varname = "national_ihme_loc_id",
  var_nat_loc_id = "national_location_id",
  keep_nat_loc_id = FALSE
)
}
\arguments{
\item{df}{[data.frame] some table with columns `location_id`}

\item{hierarchy}{[data.frame] ihme location hierarchy (get_location_metadata)}

\item{allow_self_as_parent}{[lgl: default TRUE] if TRUE, allow location_id at
level 3 to have itself as national_location_id}

\item{new_varname}{[string: default 'national_ihme_loc_id'] name of new
column to create (the national ihme_loc_id of the location_id column in
df). Used for merging on ihme_loc_id, and user may decide to retain this
column if desired.}

\item{keep_nat_loc_id}{[lgl: default FALSE] if TRUE, keep the intermediate
national_location_id column}
}
\value{
[data.frame] with new 'national_ihme_loc_id' column
}
\description{
Wrapper for `attach_national_location_id` and `attach_ihme_loc_id`.
}
\details{
Designed to fail if `national_ihme_loc_id` column is already present (since
this is a non-standard column name).
}
