% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents_of_children.R
\name{attach_parent_location_id}
\alias{attach_parent_location_id}
\title{Add `parent_location_id` column to a data.table (modified in place)}
\usage{
attach_parent_location_id(dt, hierarchy, parent_level)
}
\arguments{
\item{dt}{[data.table] some table with columns `location_id`}

\item{hierarchy}{[data.table] ihme location hierarchy (get_location_metadata)}

\item{parent_level}{[int] single parent level for all location_ids in dt}
}
\value{
[data.table] with new 'parent_id' column
}
\description{
Add `parent_location_id` column to a data.table (modified in place)
}
