% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_on_slurm_job_id.R
\name{break_for_failed_jobs}
\alias{break_for_failed_jobs}
\title{Helper function for wait_on_slurm_job_id - how do you want jobs to break and display user messages?}
\usage{
break_for_failed_jobs(
  failure_type,
  cmd_fail,
  cmd_fail_feedback,
  job_id_regex_raw,
  filter_by
)
}
\arguments{
\item{failure_type}{[chr]}

\item{cmd_fail}{[chr]}

\item{cmd_fail_feedback}{[chr]}

\item{job_id_regex_raw}{[regex]}

\item{filter_by}{[chr]}
}
\value{
[none] stop on failure
}
\description{
Helper function for wait_on_slurm_job_id - how do you want jobs to break and display user messages?
}
