% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_on_slurm_job_id.R
\name{break_for_timeout_jobs}
\alias{break_for_timeout_jobs}
\title{Helper function for wait_on_slurm_job_id - how do you want jobs to break and display user messages?}
\usage{
break_for_timeout_jobs(
  cmd_timeout,
  cmd_timeout_feedback,
  job_id_regex_raw,
  filter_by
)
}
\arguments{
\item{cmd_timeout}{[chr]}

\item{cmd_timeout_feedback}{[chr]}

\item{job_id_regex_raw}{[regex]}

\item{filter_by}{[chr]}
}
\value{
[none] stop on timeout
}
\description{
Helper function for wait_on_slurm_job_id - how do you want jobs to break and display user messages?
}
