% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{build_metadata_shell}
\alias{build_metadata_shell}
\title{Build metadata from a template}
\usage{
build_metadata_shell(
  code_root,
  jobname_filter = "^rst_ide|^vscode",
  submitline_n_char = 1000L,
  regex_to_extract = "ihme/singularity-images/rstudio/[:graph:]+",
  regex_to_ignore = "jpy",
  system_user_name = Sys.info()[["user"]],
  cluster_type = "slurm"
)
}
\arguments{
\item{code_root}{[path] path to top-level code repo folder, require: path
contains a `.git` subfolder}

\item{jobname_filter}{[character|regex] when you run `sacct/squeue -u
<username>`, what `NAME` do you want to filter for?}

\item{submitline_n_char}{[int] length of submitted command string to expect
from system (set this much longer than you'd think necessary)}

\item{regex_to_extract}{[character|regex] what string do you want to extract
after running  `sacct -j <jobid> -o submitline%xxx` using
`stringr::str_extract_all`}

\item{regex_to_ignore}{[character|regex] if your `regex_to_extract` command
finds more strings than you want, this removes all strings with}

\item{system_user_name}{[chr] user's identifier, according to the cluster}

\item{cluster_type}{[chr] methods/calls may differ by system - only 'slurm'
currently available}
}
\value{
[list] full metadata, including git info, cluster submission
  commands, and user-appended items
}
\description{
Build metadata from a template
}
