% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_brace_balance.R
\name{check_brace_balance}
\alias{check_brace_balance}
\title{Check Brace Balance in a File}
\usage{
check_brace_balance(fpath)
}
\arguments{
\item{fpath}{[chr] Path to the file to check}
}
\value{
[chr] message
}
\description{
Check Brace Balance in a File
}
\examples{
check_brace_balance("R/check_brace_balance.R")
}
