% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_of_parents.R
\name{children_of_parents}
\alias{children_of_parents}
\title{Get child locations given parent IDs from IHME hierarchies}
\usage{
children_of_parents(
  parent_loc_ids,
  path_to_parent_string,
  include_parent = TRUE
)
}
\arguments{
\item{parent_loc_ids}{Numeric vector of parent location_ids}

\item{path_to_parent_string}{Vector of strings, formatted "1,2,3". What is
the path_to_top_parent string vector?}

\item{include_parent}{Logical T/F Do you want parents included with children?
If FALSE, be careful if the parent location_ids are nested inside each
other.}
}
\value{
Logical T/F vector
}
\description{
\code{children_of_parents()} finds children from IHME Hierarchies, given a
vector of location_ids.v From a hierarchy, which children are have a given
parent ID in the path_to_top_parent column?  Returns logical T/F vector.
}
\examples{

paths <- c("1,31,32", "1,31,32,33", "1,4,5,6") # Central Asia, Armenia, China
parents <- c(6, 32) # China, Central Asia

children_of_parents(parent_loc_ids = parents, path_to_parent_string = paths, include_parent = T)
# [1] TRUE  TRUE  TRUE

children_of_parents(parent_loc_ids = parents, path_to_parent_string = paths, include_parent = F)
# [1] FALSE  TRUE FALSE

}
