% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{clean_encoding}
\alias{clean_encoding}
\title{Clean common dirty character encodings}
\usage{
clean_encoding(x)
}
\arguments{
\item{x}{[chr] vector of character strings}
}
\value{
[chr] vector of cleaned character strings
}
\description{
Not guaranteed to work in all cases - optimized for common cases
}
