% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariate_with_aggregates_hierarchy}
\alias{covariate_with_aggregates_hierarchy}
\title{covariate_with_aggregates_hierarchy}
\format{
A data.frame with 516 rows and 29 variables:
}
\usage{
covariate_with_aggregates_hierarchy
}
\description{
Example hierarchy for function development
}
\keyword{datasets}
