% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_tribble.R
\name{csv_to_tribble}
\alias{csv_to_tribble}
\title{Convert CSV to R tribble}
\usage{
csv_to_tribble(
  csv_input_path,
  output_R_path,
  obj_name,
  align_commas = TRUE,
  make_backup = FALSE
)
}
\arguments{
\item{csv_input_path}{[chr] path to input CSV file}

\item{output_R_path}{[chr] path to output R file}

\item{obj_name}{[chr] name of the R object defined in the output file}

\item{align_commas}{[lgl] whether to align commas in the tribble for better
readability}

\item{make_backup}{[lgl] whether to make a backup of the original file before
styling (`styler::tidyverse_style`) and aligning commas}
}
\value{
[invisible tibble] the data read from the CSV
}
\description{
Reads a CSV file and writes an R script that defines the same table as a
tibble using `tibble::tribble()`.
}
