% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_roots.R
\name{define_git_repo_root}
\alias{define_git_repo_root}
\title{Find the root directory of a git repository.}
\usage{
define_git_repo_root()
}
\value{
[path] path to root folder
}
\description{
Intended use is from primary launch script.
}
