% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_user_packages.R
\name{detach_user_packages}
\alias{detach_user_packages}
\title{Detach all user-loaded packages}
\usage{
detach_user_packages()
}
\value{
[nothing]
}
\description{
Start with a clean environment, devoid of packages that may cause namespace
function conflicts (e.g. plyr & dplyr).  Runs in a loop to deal with
package dependencies while unloading.
}
