% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dice.R
\name{dice}
\alias{dice}
\title{Roll dice}
\usage{
dice(n_dice, n_sides, best_of_n = 1, method = c("max", "min"), verbose = FALSE)
}
\arguments{
\item{n_dice}{e.g. 3 of 3d6}

\item{n_sides}{e.g 6 of 3d6}

\item{best_of_n}{if rolling 3d6 n times, choose the highest/lowest sum from n rolls}

\item{method}{"max" (advantage) or "min" (disadvantage)}
}
\value{
a barplot showing the distribution of sums
}
\description{
Roll dice
}
\examples{
dice(3, 6, 2)
}
