% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_inequal_prob}
\alias{draws_inequal_prob}
\title{Calculate probability that draws of comp_var for comp_vec[1] are <, <=, >, or >=
draws of comp_var for comp_vec[2].}
\usage{
draws_inequal_prob(DT, comp_var, comp_vec, operator, by_vars)
}
\arguments{
\item{DT}{[data.table] a table of (long) draws}

\item{comp_var}{[chr] variable name to compare - e.g. "year_id"}

\item{comp_vec}{[chr] vector of 2 variable levels to compare - e.g. c("1990", "2020")}

\item{operator}{[chr: {"lt", "lte", "gt", "gte"}] operator for comparison -
translates to <, <=, >, >= internally}

\item{by_vars}{[chr] variable names to group by when calculating probability
- e.g. location_id}
}
\value{
[data.table] a table of probabilities by `by_vars`
}
\description{
The comparison is done at the draw level after pivoting wide by comp_var.
}
