% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_var_to_wide}
\alias{draws_var_to_wide}
\title{Pivot long draws wide by some variable}
\usage{
draws_var_to_wide(DT, varname, var_vec = NULL)
}
\arguments{
\item{DT}{[data.table] input draws}

\item{var_vec}{[chr] vector of variable levels to pivot wide}
}
\value{
[data.table] draws in wide format
}
\description{
Pivot long draws wide by some variable
}
