% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_wide_to_long}
\alias{draws_wide_to_long}
\title{Transform draws from wide to long format creating a `draw_id` column}
\usage{
draws_wide_to_long(
  DT,
  id_varnames = find_id_varnames(DT, verbose = FALSE),
  verbose = FALSE,
  chk_square = TRUE
)
}
\arguments{
\item{DT}{[data.table] input draws}

\item{id_varnames}{[character] columns to keep as is}

\item{verbose}{[lgl: default FALSE] print debug messages?}

\item{chk_square}{[lgl: default TRUE] assert that DT is square before transforming?}
}
\value{
[data.frame] draws in long format
}
\description{
Transform draws from wide to long format creating a `draw_id` column
}
