% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_year_diff}
\alias{draws_year_diff}
\title{Compare two-years diffs at the draw level}
\usage{
draws_year_diff(
  DT,
  yr_vec,
  id_varnames = find_id_varnames(DT, verbose = FALSE)
)
}
\arguments{
\item{DT}{[data.table] a table of (long) draws}

\item{yr_vec}{[int] 2 years to compare}

\item{id_varnames}{[character] columns to keep as id columns}
}
\value{
[data.table] a table of mean and 95% CI for the difference between
  the two years, by id_varnames
}
\description{
Calculate year 2 minus year 1.  Helper function that wraps other draw
transformations,  useful for publications.
}
