% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_year_prob}
\alias{draws_year_prob}
\title{Calculate probability that draws in year 2 are greater than (or less than)
draws in year 1.}
\usage{
draws_year_prob(DT, yr_vec, operator, by_vars)
}
\arguments{
\item{DT}{[data.table] a table of (long) draws}

\item{yr_vec}{[int] 2 years to compare (function sorts them internally)}

\item{operator}{[chr: {"lt", "lte", "gt", "gte"}] operator for comparison -
translates to <, <=, >, >= internally}

\item{by_vars}{[chr] variable names to group by when calculating probability
- e.g. location_id}
}
\value{
[data.table] a table of probabilities by `by_vars`
}
\description{
Comparison is ALWAYS year 2 vs year 1 (e.g. year 2 < year 1)
}
