% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{extract_command_string}
\alias{extract_command_string}
\title{Search a cluster submitted command string for some pattern}
\usage{
extract_command_string(
  submit_command_text,
  regex_to_extract,
  regex_to_ignore = NULL
)
}
\arguments{
\item{submit_command_text}{[chr] the result of calling "sacct -j <INT> -o
submitline\%<INT>"}

\item{regex_to_extract}{[regex] pattern to extract from `submit_command_text`}

\item{regex_to_ignore}{[regex] ignore strings found with this pattern}
}
\value{
[chr] vector of strings extracting/ignoring as requested
}
\description{
Intended to pull Rstudio image information
- Currently only supports one valid string found per submission command
}
