% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{extract_cores}
\alias{extract_cores}
\title{Extract number of interactive user cores}
\usage{
extract_cores()
}
\value{
[int] number of available cores for multithreading
- if user has more than one interactive session, defaults to 1
}
\description{
Extract number of interactive user cores
}
