% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{extract_script_path}
\alias{extract_script_path}
\title{Extract the path to the current script}
\usage{
extract_script_path(code_root)
}
\arguments{
\item{code_root}{[chr] Path to top-level code repo folder, require: path}
}
\value{
[chr] Path to the current script, relative to the code_root, NA if
  unable to find
}
\description{
Extract the path to the current script
}
\examples{
extract_script_path(code_root = getwd())
}
