% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{extract_sessionInfo}
\alias{extract_sessionInfo}
\title{Extract key pieces of session info to avoid metadata bloat}
\usage{
extract_sessionInfo(
  fields = c("Package", "Title", "Version", "Depends", "Imports", "Date/Publication")
)
}
\arguments{
\item{fields}{[chr] package info fields to extract from sessionInfo}
}
\value{
[list] sessionInfo with only the specified fields
}
\description{
Extract key pieces of session info to avoid metadata bloat
}
\examples{
extract_sessionInfo()
}
