% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fh_small_area_hierarchy}
\alias{fh_small_area_hierarchy}
\title{fh_small_area_hierarchy}
\format{
A data.frame with 305 rows and 29 variables:
}
\usage{
fh_small_area_hierarchy
}
\description{
Example hierarchy for function development
}
\keyword{datasets}
