% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{find_draws_varnames}
\alias{find_draws_varnames}
\title{Find draw variable names in a data.table}
\usage{
find_draws_varnames(DT, draws_rgx = PERD_regex())
}
\arguments{
\item{DT}{[data.table]}

\item{draws_rgx}{[chr] regex to identify draw columns + point estimate}
}
\value{
[chr] vector of variable names
}
\description{
Find draw variable names in a data.table
}
