% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{find_file_extension}
\alias{find_file_extension}
\title{Find the file extension from a full path}
\usage{
find_file_extension(f_path)
}
\arguments{
\item{f_path}{[path]}
}
\value{
[chr] a file extension
}
\description{
Find the file extension from a full path
}
