% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{find_id_varnames}
\alias{find_id_varnames}
\title{Util to find a potential set of id_varnames for a data.table full of draws}
\usage{
find_id_varnames(DT, additions = NULL, removals = "value", verbose = TRUE)
}
\arguments{
\item{DT}{[data.frame]}
}
\value{
[chr] vector of variable names
}
\description{
Interactive helper
}
