% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{fmt_magnitude}
\alias{fmt_magnitude}
\title{Format magnitude}
\usage{
fmt_magnitude(x, digits = 1, nsmall = 1, mag = "m", allow_thousands = FALSE)
}
\arguments{
\item{x}{[num] numeric vector}

\item{digits}{[int] passed to `round()`}

\item{nsmall}{[int] passed to `format()`}

\item{mag}{[chr c("b", "m", "t")] magnitude (billion, million, thousand)}
}
\value{
[chr] formatted string
}
\description{
Caution - thousands magnitude is not Lancet compliant
}
\examples{
fmt_magnitude(123456789) # "123.5 million"
}
