% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{format_journal_clu}
\alias{format_journal_clu}
\title{Format central/lower/upper value triplets for journal presentation}
\usage{
format_journal_clu(
  central,
  lower,
  upper,
  d_type,
  digits_round_prop = 1L,
  digits_sigfig_count = 3L,
  nsmall = 1L,
  decimal.mark = ".",
  negative_sign = "-",
  big.mark_count = ",",
  mean_neg_text = "a decrease of ",
  UI_only = FALSE,
  assert_clu_relationships = TRUE,
  is_lancet = FALSE
)
}
\arguments{
\item{central}{[num] central/point_estimate value vector}

\item{lower}{[num] lower bound vector}

\item{upper}{[num] upper bound vector}

\item{d_type}{[chr {prop, pp, count}] data type - proportion, percentage
point or count}

\item{digits_round_prop}{[int] number of digits to round proportions/PP}

\item{digits_sigfig_count}{[int] number of significant digits for counts}

\item{nsmall}{[int] number of digits after the decimal point}

\item{decimal.mark}{[chr] decimal mark}

\item{negative_sign}{[chr] negative sign}

\item{big.mark_count}{[chr] big mark for counts}

\item{mean_neg_text}{[chr: default "a decrease of "] text to prepend if all
central/upper/lower triplet values are negative}

\item{UI_only}{[lgl] return only the UI?}

\item{assert_clu_relationships}{[lgl] enforce correct relationship between
central/upper/lower values? This should _ALWAYS_ be `TRUE` _UNLESS_ you
only care about the central difference.}

\item{is_lancet}{[lgl] apply Lancet-specific formatting for counts 9999 and smaller}
}
\value{
[chr] formatted string vector
}
\description{
Defaults are generic.  This function allows special formtting marks to be
applied by journal. Use `format_lancet_clu()` for Lancet-specific formatting.
Use `format_nature_clu()` for Nature-specific formatting.
}
\details{
Takes three vectors as main arguments for data.table-friendly vectorization.

`central` could be mean/median/point_estimate

Transform c(central = 0.994, lower = 0.984, upper = 0.998) to "99.4%
(98.4–99.8)"

Accounts for negative values, and UIs that cross zero.  Checks if
central/lower/upper values are in the correct order.
- https://hub.ihme.washington.edu/spaces/SGT/pages/45675035/WRITING+STYLE+GUIDE#WRITINGSTYLEGUIDE-GBDStyleGuideGBDStyleGuide
}
