% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{format_journal_dt}
\alias{format_journal_dt}
\title{Return a table with formatted central/lower/upper}
\usage{
format_journal_dt(
  dt,
  d_type,
  central_var = "mean",
  lower_var = "lower",
  upper_var = "upper",
  remove_clu = TRUE,
  assert_clu_relationships = TRUE,
  digits_round_prop = 1L,
  digits_sigfig_count = 3L,
  nsmall = 1L,
  decimal.mark = ".",
  negative_sign = "-",
  big.mark_count = "",
  mean_neg_text = "a decrease of ",
  UI_only = FALSE,
  is_lancet = FALSE
)
}
\arguments{
\item{dt}{[data.table]}

\item{d_type}{[chr {'prop', 'pp', or 'count'}] a single data type}

\item{central_var}{[chr: default 'mean'] name of central tendency variable}

\item{lower_var}{[chr: default 'lower'] name of lower bound variable}

\item{upper_var}{[chr: default 'upper'] name of upper bound variable}

\item{remove_clu}{[lgl] remove central/lower/upper variables after
formatting?}

\item{assert_clu_relationships}{[lgl] enforce correct relationship between
central/upper/lower values? This should _ALWAYS_ be `TRUE` _UNLESS_ you
only care about the central difference.}

\item{digits_round_prop}{[int] passed to `round()` for proportions}

\item{digits_sigfig_count}{[int] passed to `signif()` for counts}

\item{nsmall}{[int] passed to `format()`}

\item{decimal.mark}{[chr] decimal mark passed to `format()`}

\item{negative_sign}{[chr] negative sign}

\item{big.mark_count}{[chr] big mark for counts passed to `format()`}

\item{mean_neg_text}{[chr: default "a decrease of "] text to prepend if all
central/upper/lower triplet values are negative}

\item{UI_only}{[lgl] return only the UI?}

\item{is_lancet}{[lgl] apply Lancet-specific formatting for counts 9999 and
smaller}
}
\value{
[data.table] copy of input data.table with new 'clu_fmt' column
}
\description{
Assumes a single data-type (d_type) for the whole table (e.g. 'prop', 'pp',
'count')
}
