% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{format_lancet_clu}
\alias{format_lancet_clu}
\title{Lancet Format central/lower/upper data triplets (vectorized)}
\usage{
format_lancet_clu(
  central,
  lower,
  upper,
  d_type,
  digits_round_prop = 1L,
  digits_sigfig_count = 3L,
  nsmall = 1L,
  decimal.mark = mid_dot(),
  negative_sign = en_dash(),
  big.mark_count = thin_space(),
  mean_neg_text = "a decrease of ",
  UI_only = FALSE,
  assert_clu_relationships = TRUE,
  is_lancet = TRUE
)
}
\arguments{
\item{central}{[num] central/point_estimate value vector}

\item{lower}{[num] lower bound vector}

\item{upper}{[num] upper bound vector}

\item{d_type}{[chr {prop, pp, count}] data type - proportion,
percentage point or count}

\item{digits_round_prop}{[int] number of digits to round proportions/PP}

\item{digits_sigfig_count}{[int] number of significant digits for counts}

\item{nsmall}{[int] number of digits after the decimal point}

\item{decimal.mark}{[chr] decimal mark - Lancet specifies mid-dot}

\item{negative_sign}{[chr] negative sign}

\item{big.mark_count}{[chr] big mark for counts - Lancet specifies narrow
space for counts between 9999 and 1 million}

\item{mean_neg_text}{[chr: default "a decrease of "] text to prepend if central is negative
e.g. Lancet prefers "a decrease of 99·4\% (98·4–99·8)"}

\item{UI_only}{[lgl] return only the UI?}

\item{assert_clu_relationships}{[lgl] enforce correct relationship between central/upper/lower
values? This should _ALWAYS_ be `TRUE` _UNLESS_ you only care about the
central difference.}
}
\value{
[chr] formatted string vector
}
\description{
Takes three vectors as main arguments for data.table-friendly vectorization.
}
\details{
`central` could be mean/median/point_estimate

Transform c(central = 0.994, lower = 0.984, upper = 0.998) to "99.4% (98.4–99.8)"

Accounts for negative values, and UIs that cross zero.  Checks if
central/lower/upper values are in the correct order.
- https://hub.ihme.washington.edu/spaces/SGT/pages/45675035/WRITING+STYLE+GUIDE#WRITINGSTYLEGUIDE-GBDStyleGuideGBDStyleGuide
}
\examples{
format_lancet_clu(central = 0.994, lower = 0.984, upper = 0.998, d_type = "prop") # "99.4\% (98.4–99.8)"
format_lancet_clu(central = c(0.994, 0.994), lower = c(0.984, 0.984), upper = c(0.998, 0.998), d_type = "prop") # "99·4\% (98·4–99·8)" "99·4\% (98·4–99·8)"
format_lancet_clu(central = c(0.994, 0.994), lower = c(-0.15, 0.984), upper = c(0.998, 0.998), d_type = "prop") # "1·0\% (–0·1 to 1·0)" "1·0\% (1·0–1·0)"
format_lancet_clu(central = c(-0.05, 0.994), lower = c(-0.15, 0.984), upper = c(0.998, 0.998), d_type = "pp") # "1·0 pp (–0·1 to 1·0)" "1·0 pp (1·0–1·0)"
format_lancet_clu(central = rep(2, 2), lower = rep(.5, 2), upper = rep(3, 2), d_type = "count")
format_lancet_clu(central = rep(2e6, 2), lower = rep(.5e6, 2), upper = rep(3e6, 2), d_type = "count")
format_lancet_clu(central = c(-0.994, -0.994), upper = c(-0.984, -0.984), lower = c(-0.998, -0.998), d_type = "prop",  digits_round_prop = 4)
format_lancet_clu(central = c(-0.994, -0.994), upper = c(-0.984, -0.984), lower = c(-0.998, -0.998), d_type = "prop",  digits_round_prop = 4, UI_only = T)
format_lancet_clu(central = c(0.994, 0.994), lower = c(0.984, 0.984), upper = c(0.998), d_type = "prop") # fail
format_lancet_clu(central = c(0.994, 0.999), lower = c(0.984, 0.984), upper = c(0.998, 0.998), d_type = "prop") # fail
format_lancet_clu(central = c(0.994, 0.994), lower = c(0.984, 0.984), upper = c(0.998, 0.998), d_type = "propeller") # fail
}
