% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{format_means_dt}
\alias{format_means_dt}
\title{Format means in data.table by data type for presentation.}
\usage{
format_means_dt(
  DT,
  dtype,
  central_varname = "mean",
  scalar = 100,
  digits = 1,
  decimal.mark = ".",
  big.mark = ","
)
}
\arguments{
\item{DT}{[data.table] input data.table with one or more 'mean_' columns}

\item{dtype}{[chr {'prop', 'pp', or 'count'}] a single data type}

\item{central_varname}{[chr: default 'mean'] prefix of mean variable names to
format.  Implemented as e.g. "^mean[_]*" to capture 'mean', 'mean_1990',
'mean_2000', etc.}

\item{scalar}{[numeric: default 100] scalar to multiply mean values by}

\item{digits}{[integer: default 1] number of digits to round to}
}
\value{
[data.table] copy of input data.table with formatted mean column(s)
}
\description{
TODO 2025-10-30 - not yet configured with set_magnitude
}
\details{
Multiply one or more 'mean_' columns by a scalar and round to specified
digits based on a data type.

Pairs nicely with `means_year_compare()`

Data types:
- 'prop'  - proportion (e.g. 0.1234 -> 12.3%)
- 'pp'    - percentage point (e.g. 0.1234 -> 12.3 pp)
- 'count' - count (e.g. 1234.56 -> 1235)
}
\examples{
DT <- data.table::data.table(
  location_id = c(1, 2, 3)
  , mean_1990 = c(0.1234, 0.2345, 0.3456)
  , mean_2000 = c(0.2234, 0.3345, 0.4456)
 )

format_means_dt(DT, dtype = "prop")
}
