% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{format_nature_clu}
\alias{format_nature_clu}
\title{Format central/lower/upper sets to Nature CLU format (vectorized)}
\usage{
format_nature_clu(
  central,
  lower,
  upper,
  d_type,
  digits_round_prop = 1L,
  digits_sigfig_count = 3L,
  nsmall = 1L,
  decimal.mark = ".",
  negative_sign = "-",
  big.mark_count = ",",
  mean_neg_text = "a decrease of ",
  UI_only = FALSE,
  UI_text = "95\% uncertainty interval, ",
  assert_clu_relationships = TRUE
)
}
\arguments{
\item{central}{[num] central/point_estimate value vector}

\item{lower}{[num] lower bound vector}

\item{upper}{[num] upper bound vector}

\item{d_type}{[chr {prop, pp, count}] data type - proportion, percentage
point or count}

\item{digits_round_prop}{[int] number of digits to round proportions/PP}

\item{digits_sigfig_count}{[int] number of significant digits for counts}

\item{nsmall}{[int] number of digits after the decimal point}

\item{decimal.mark}{[chr] decimal mark}

\item{negative_sign}{[chr] negative sign}

\item{big.mark_count}{[chr] big mark for counts}

\item{mean_neg_text}{[chr: default "a decrease of "] text to prepend if all
central/upper/lower triplet values are negative}

\item{UI_only}{[lgl: default FALSE] return only the UI?}

\item{UI_text}{[chr: default "95% uncertainty interval, "] text to appear
before values inside UI parentheses}

\item{assert_clu_relationships}{[lgl] enforce correct relationship between
central/upper/lower values? This should _ALWAYS_ be `TRUE` _UNLESS_ you
only care about the central difference.}
}
\value{
[chr] formatted string vector
}
\description{
Format central/lower/upper sets to Nature CLU format (vectorized)
}
