% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{format_nature_dt}
\alias{format_nature_dt}
\title{Title}
\usage{
format_nature_dt(
  dt,
  d_type,
  central_var = "mean",
  lower_var = "lower",
  upper_var = "upper",
  remove_clu = TRUE,
  assert_clu_relationships = TRUE,
  digits_round_prop = 1L,
  digits_sigfig_count = 3L,
  nsmall = 1L,
  decimal.mark = ".",
  negative_sign = "-",
  big.mark_count = ",",
  mean_neg_text = "a decrease of ",
  UI_only = FALSE,
  UI_text = "",
  is_lancet = FALSE
)
}
\arguments{
\item{dt}{[data.table]}

\item{d_type}{[chr {'prop', 'pp', or 'count'}] a single data type}

\item{central_var}{[chr: default 'mean'] name of central tendency variable}

\item{lower_var}{[chr: default 'lower'] name of lower bound variable}

\item{upper_var}{[chr: default 'upper'] name of upper bound variable}

\item{remove_clu}{[lgl: default TRUE] remove central/lower/upper variables
after}

\item{assert_clu_relationships}{[lgl: default TRUE] enforce correct
relationship between}

\item{digits_round_prop}{[int: default 1L] passed to `round()` for
proportions}

\item{digits_sigfig_count}{[int: default 3L] passed to `signif()` for counts}

\item{nsmall}{[int: default 1L] passed to `format()`}

\item{decimal.mark}{[chr: default "."] decimal mark passed to `format()`}

\item{negative_sign}{[chr: default "-"] negative sign}

\item{big.mark_count}{[chr: default ","] big mark for counts passed to
`format()`}

\item{mean_neg_text}{[chr: default "a decrease of "] text to prepend if}

\item{UI_only}{[lgl: default FALSE] return only the UI?}

\item{UI_text}{[chr: default "95% uncertainty interval, "] text to appear
before values inside UI parentheses}

\item{is_lancet}{[lgl: default FALSE] apply Lancet-specific formatting for
counts 9999 and smaller}
}
\value{
[data.table] copy of input data.table with new 'clu_fmt' column
}
\description{
Title
}
