% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{format_oxford_comma}
\alias{format_oxford_comma}
\title{Format vector of items with Oxford comma}
\usage{
format_oxford_comma(vec)
}
\arguments{
\item{vec}{[any] vector of items to format}
}
\value{
[chr] formatted string with Oxford comma
}
\description{
Format vector of items with Oxford comma
}
\examples{
format_oxford_comma(1:2) # "1 and 2"
format_oxford_comma(1:3) # "1, 2, and 3"
}
