% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_lancet.R
\name{fround}
\alias{fround}
\title{Format and round}
\usage{
fround(x, digits = 1L, nsmall = 1L, decimal.mark = mid_dot())
}
\arguments{
\item{x}{[num] numeric vector}

\item{digits}{[integer] passed to `round()`}

\item{nsmall}{[integer] passed to `format()`}
}
\value{
[chr] formatted string
}
\description{
Unaware of data-type
}
\examples{
fround(0.123456789) # "0·1"
fround(0.123456789, digits = 3) # "0·123"
fround(0.123456789, digits = 3, nsmall = 4) # "0·1230"
}
