% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{fround_dtype}
\alias{fround_dtype}
\title{Format and round with data-type suffix}
\usage{
fround_dtype(x, d_type = "prop", digits = 1L, nsmall = 1L, decimal.mark = ".")
}
\arguments{
\item{x}{[num] numeric value}

\item{d_type}{[chr {'prop', 'pp', or 'count'}] data type - proportion,
percentage point or count}

\item{digits}{[integer: default 1L] passed to `round()`}

\item{nsmall}{[integer: default 1L] passed to `format()`}

\item{decimal.mark}{[chr: default "."] decimal mark passed to `format()`}
}
\value{
[chr] formatted string
}
\description{
Format and round with data-type suffix
}
\examples{
fround_dtype(0.123456789) # "0.1\%"
fround_dtype(0.123456789, 'pp', 3, 4) # "0.1230 pp"
fround_dtype(c(55.8346, 123.456789), 'count', 3, 4, ".") # "55.8350"  "123.4570"
}
