% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{fround_mag_clu}
\alias{fround_mag_clu}
\title{Format and round central/lower/upper value sets by magnitude _without_ units}
\usage{
fround_mag_clu(
  clu,
  d_type,
  mag_list = set_magnitude(clu[1]),
  digits_round_prop = 1L,
  digits_sigfig_count = 3L,
  nsmall = 1L,
  decimal.mark = ".",
  negative_sign = "-",
  big.mark_count = ",",
  is_lancet = FALSE
)
}
\arguments{
\item{clu}{[num] a numeric vector of central/lower/upper}

\item{d_type}{[chr {'prop', 'pp', or 'count'}] data type - proportion,
percentage point or count}

\item{mag_list}{[named list] output from `set_magnitude()` - must be based on
**central** value of a central/lower/upper set - central _and_ all UI
values inherit the same scale as the central tendency.}

\item{digits_round_prop}{[int: default 1L] passed to `round()` for
proportions}

\item{digits_sigfig_count}{[int: default 3L] passed to `signif()` for counts}

\item{nsmall}{[int: default 1L] passed to `format()`}

\item{decimal.mark}{[chr: default "."] decimal mark passed to `format()`}

\item{negative_sign}{[chr: default "-"] negative sign}

\item{big.mark_count}{[chr: default ","] big mark for counts passed to
`format()`}
}
\value{
[chr] formatted string (vectorized)
}
\description{
`central` could be mean/median/point_estimate. d_type is required (count data
requires nuanced logic), but labels are not returned.
}
\details{
Format and round without unit labeling
- Use `format_lancet_clu()` for unit labels

Defaults are for lancet labels.

Current IHME standards:
- https://hub.ihme.washington.edu/spaces/SC/pages/123833253/Rounding+standards
- Rates and Percentages - to one decimal place
- Counts                - to 3 sig figs
}
\examples{
fround_mag_clu(clu = c(central = 0.2, lower = 0.1, upper = 0.3), d_type = "prop") # "0.2" "0.1" "0.3"
fround_mag_clu(clu = c(central = 0.2, lower = -0.1, upper = 0.3), d_type = "pp") # "0.2" "–0.1" "0.3"
fround_mag_clu(clu = c(central = 95e6, lower = 89e6, upper = 101e6), d_type = "count") # "95.0"  "89.0"   "101"
fround_mag_clu(clu = c(central = 95e6, lower = 96e6, upper = 97e6), d_type = "count") # "95.0"  "96.0"  "97.0"
}
