% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbd_analysis_hierarchy}
\alias{gbd_analysis_hierarchy}
\title{gbd_analysis_hierarchy"}
\format{
A data.frame with 1126 rows and 31 variables:
}
\usage{
gbd_analysis_hierarchy
}
\description{
Example hierarchy for function development
}
\keyword{datasets}
