% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{get_latest_output_dir}
\alias{get_latest_output_dir}
\title{Find the latest output directory with format YYYY_MM_DD.VV}
\usage{
get_latest_output_dir(root, suffix = "")
}
\arguments{
\item{root}{[chr] path to root of output results}

\item{suffix}{[chr] a suffix to append after the incremented version number index}
}
\value{
[chr] path to latest output directory
}
\description{
Find the latest output directory with format YYYY_MM_DD.VV
}
\examples{
get_latest_output_dir("tests/testthat/fixtures/versioned-dirs/nested/1999_09_09") # expect "tests/testthat/fixtures/versioned-dirs/nested/1999_09_09/1999_09_09.02"
}
