% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_output_dir.R
\name{get_new_output_dir}
\alias{get_new_output_dir}
\title{Increment a new output folder date-version}
\usage{
get_new_output_dir(root, date, suffix = "")
}
\arguments{
\item{root}{[chr] path to root of output results}

\item{date}{[chr] character date in form of "YYYY_MM_DD" or "today". "today" will be interpreted as today's date.}

\item{suffix}{[chr] a suffix to append after the incremented version number index}
}
\value{
[chr] path to new output directory of the form "/<root>/YYYY_MM_DD.VV"
}
\description{
Get a new directory path, but don't make it
}
\examples{
get_new_output_dir(root = tempdir(), date = "today")
}
