% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{get_new_output_dv}
\alias{get_new_output_dv}
\title{Increment a new output folder date-version}
\usage{
get_new_output_dv(root, date = "today", suffix = "")
}
\arguments{
\item{root}{[chr] path to root of output results}

\item{date}{[chr] character date in form of "YYYY_MM_DD" or "today". "today" will be interpreted as today's date.}

\item{suffix}{[chr] a suffix to append after the incremented version number index}
}
\value{
[dbl] new output version of the form "YYYY_MM_DD.VV"
}
\description{
Return on the date-version, not the full path.  Does not create a folder.
}
\examples{
get_new_output_dv(root = tempdir(), date = "today")
}
