% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_from_path.R
\name{get_version_from_path}
\alias{get_version_from_path}
\title{Get YYYY_MM_DD.VV version from an output path}
\usage{
get_version_from_path(x)
}
\arguments{
\item{x}{[file path] versioned output folder}
}
\value{
[char] "YYYY_MM_DD.VV" folder name
}
\description{
Really just retuns the last element of a string split on "/".
This function currently has no validation or safety rails
}
