% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_helpers.R
\name{git_diff}
\alias{git_diff}
\title{Record uncommitted git changes}
\usage{
git_diff(CODE_ROOT)
}
\arguments{
\item{CODE_ROOT}{[path] path to code root containing a .git folder}
}
\value{
[character] succinct git diff vector length 1 with newlines for easy console printing
}
\description{
Record uncommitted git changes
}
