% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{increment_file_version}
\alias{increment_file_version}
\title{If a file exists, get a new path with `v1`, `v2`, etc. appended}
\usage{
increment_file_version(outpath)
}
\arguments{
\item{outpath}{[chr] full path to file}
}
\value{
[chr] new path with version number appended (if necessary)
}
\description{
If file does not exist, return the original path
}
\examples{
dir.create(tempdir(), recursive = TRUE, showWarnings = FALSE)
fname_old <- "test_file.csv"
file.create(file.path(tempdir(), fname_old))
list.files(tempdir()) # [1] "test_file.csv"
fname_new <- increment_file_version(file.path(tempdir(), "test_file.csv"))
file.create(fname_new)
list.files(tempdir()) # 1] "test_file_v1.csv" "test_file.csv"
file.remove(c(file.path(tempdir(), fname_old), fname_new))
}
