% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_x.R
\name{is_cli_null}
\alias{is_cli_null}
\title{Check if an element would be empty when passed as a CLI argument}
\usage{
is_cli_null(x)
}
\arguments{
\item{x}{[vec] some scalar element}
}
\value{
[lgl] TRUE if the element would be passed to the CLI as '', else FALSE
}
\description{
Check if an element would be empty when passed as a CLI argument
}
\examples{
is_cli_null(c("A", 5, ''))
}
