% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_x.R
\name{is_empty}
\alias{is_empty}
\title{Check that a vector/element is not empty}
\usage{
is_empty(x, check_whitespace = TRUE)
}
\arguments{
\item{x}{[vec] some scalar element}

\item{check_whitespace}{[lgl] check for all-whitespace strings?}
}
\value{
[lgl]
}
\description{
Check for is.null & length 0. Vectorized check for is.na, empty string (""),
optional check for whitespace only string.
}
