% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{is_sequential_int_vec}
\alias{is_sequential_int_vec}
\title{Are all elements of a vector contiguous integers?}
\usage{
is_sequential_int_vec(x)
}
\arguments{
\item{x}{[int] an integer vector
::}
}
\value{
[lgl] TRUE if all elements are contiguous integers
}
\description{
Useful for checking if a vector is a valid array task list.
- uses user-submitted order
}
\examples{
is_contiguous_int_vec(1L:10L) # TRUE
is_contiguous_int_vec(c(1L, 3L, 2L)) # FALSE
}
