% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_iso_to_ihme_loc_id.R
\name{iso_map_France_level_3}
\alias{iso_map_France_level_3}
\title{Map ISO3 codes for France level 3 (Metropolitan + overseas territories)}
\usage{
iso_map_France_level_3()
}
\value{
[chr] named vector mapping certain `ihme_loc_id` values to their ISO3 codes
}
\description{
Map ISO3 codes for France level 3 (Metropolitan + overseas territories)
}
\examples{
iso_map_France_level_3()
}
