% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{make_new_output_dir}
\alias{make_new_output_dir}
\title{Get output directory for results to save in}
\usage{
make_new_output_dir(root, date)
}
\arguments{
\item{root}{[chr] path to root of output results}

\item{date}{[chr] character date in form of "YYYY_MM_DD" or "today". "today" will be interpreted as today's date.}
}
\value{
[chr] path to new output directory
}
\description{
Returns an appropriate path to save results in, creating it if necessary.
}
\examples{
\dontrun{
make_new_output_dir("my/root/folder", date = "today")
}
}
