% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_iso_to_ihme_loc_id.R
\name{map_iso_to_ihme_loc_id}
\alias{map_iso_to_ihme_loc_id}
\title{Map certain `ihme_loc_id` values to their ISO3 codes}
\usage{
map_iso_to_ihme_loc_id(
  dt,
  iso_varname = "ihme_loc_id",
  loc_subsets = NULL,
  subset_exact = NULL,
  direction = "ltr"
)
}
\arguments{
\item{dt}{[data.table] of data with `ihme_loc_id` column}

\item{iso_varname}{[chr: default "ihme_loc_id"] variable name in `dt` for
`ihme_loc_id`}

\item{loc_subsets}{[chr or NULL: default NULL] subsets of mappings to apply,
NULL means all}

\item{subset_exact}{[chr or NULL: default NULL] if not NULL, only apply
mappings for these exact `ihme_loc_id` or `ISO3` values}

\item{direction}{[chr: default "ltr"] direction of mapping, "ltr" means
`ihme_loc_id` to `ISO3`, "rtl" means `ISO3` to `ihme_loc_id`}
}
\value{
[data.table] input `dt` with `ihme_loc_id` values replaced by ISO3
  codes where applicable
}
\description{
User control of variable name for `ihme_loc_id` in case it differs from the
default.
}
\details{
Started to handle Metropolitan France + overseas DROMCOM territories, may
expand to other location subsets later.
}
\examples{
dt <- data.table::data.table(ihme_loc_id = c("FRA", "USA", "GLP", "CAN"), value = 1:4)
dt_mapped <- map_iso_to_ihme_loc_id(data.table::copy(dt), loc_subsets = c("France"))
print(dt)
print(dt_mapped)
}
