% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_transforms.R
\name{means_year_compare}
\alias{means_year_compare}
\title{Yearly Mean Comparison}
\usage{
means_year_compare(
  DT,
  yr_vec,
  central_varname = "mean",
  id_varnams = find_id_varnames(DT, verbose = FALSE)
)
}
\arguments{
\item{DT}{[data.table] of long means with a 'year_id' column}

\item{yr_vec}{[integer] vector of years to compare (at least 2)}

\item{central_varname}{[string: default 'mean'] name of the central tendency
variable column}

\item{id_varnams}{[chr: default `find_id_varnames()`] id variable names that
describe the squareness of DT}
}
\value{
[data.table] input `DT` pivoted wider with mean columns for each
  year
}
\description{
Pivot a table of long means wider by selected years (2 or more).  Works for any 'central
tendency' variable (e.g., mean, median, point_estimate).
}
\examples{
DT <- data.table::data.table(
  location_id = rep(1:3, each = 4)
  , year_id    = rep(2000:2003, times = 3)
  , mean       = rnorm(12, mean = 100, sd = 10)
)
means_year_compare(
  DT = DT
  , yr_vec = c(2000,2003)
  , central_varname = "mean"
  , id_varnams = c("location_id", "year_id")
)
}
