% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_df_list.R
\name{merge_df_list}
\alias{merge_df_list}
\title{Merge a list of same-structured data.frames by common columns}
\usage{
merge_df_list(df_list, merge_by)
}
\arguments{
\item{df_list}{[list] of data.frames to merge}

\item{merge_by}{[chr] vector of variable names to merge by}
}
\value{
[data.frame] merged data.frame
}
\description{
For example merge a list of five data-frame into one, assuming they all share
a common structure and can be merged by the same columns.
}
\examples{
df1 <- data.frame(id = 1:3, value1 = c(10, 20, 30))
df2 <- data.frame(id = 2:4, value2 = c(15, 25, 35))
df3 <- data.frame(id = 3:5, value3 = c(12, 22, 32))
df_list <- list(df1, df2, df3)
merge_df_list(df_list, merge_by = "id")
}
