% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modeling_hierarchy}
\alias{modeling_hierarchy}
\title{modeling_hierarchy}
\format{
A data.frame with 465 rows and 29 variables:
}
\usage{
modeling_hierarchy
}
\description{
Example hierarchy for function development
}
\keyword{datasets}
