% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msg.R
\name{msg_multiline}
\alias{msg_multiline}
\title{Print a message with multiple lines}
\usage{
msg_multiline(x)
}
\arguments{
\item{x}{[list] list or data.frame to print}
}
\value{
[none] message output
}
\description{
Print a message with multiple lines
}
\examples{
msg_multiline(head(mtcars, 2))
                  mpg cyl disp  hp drat    wt  qsec vs am gear carb
Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
}
