% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_prt.R
\name{msg_prt}
\alias{msg_prt}
\title{Send a user-message to both stdout and stderr}
\usage{
msg_prt(string = "No message supplied")
}
\arguments{
\item{string}{[character] message to the pipeline's user. Require: length = 1}
}
\value{
[stdout/stderr] user message to both stdout and stderr
}
\description{
When using `sink()` on a pipeline, some context may be more useful in stdout
or stderr, but it's difficult to know when.  This wrapper prints a message to
both locations.
}
