% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msg.R
\name{msg_section}
\alias{msg_section}
\title{StdErr section header with nice spacing}
\usage{
msg_section(..., section = "", newlines = 2L)
}
\arguments{
\item{...}{[chr] passed to message()}

\item{section}{[chr] a section header before the message}

\item{newlines}{[int: default TRUE] number of newlines to add before and
after message}
}
\value{
[none] side effect - displays to std_err
}
\description{
StdErr section header with nice spacing
}
\examples{
msg_section("/my/file/path", section = "my path")
}
