% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msg.R
\name{msg_toc}
\alias{msg_toc}
\title{Message tictoc toc time with prefix}
\usage{
msg_toc(prefix = " -- ")
}
\arguments{
\item{prefix}{[chr: default " -- "] prefix to the message}
}
\value{
[none]
}
\description{
Message tictoc toc time with prefix
}
\examples{
msg_tic()
msg_toc()
}
\seealso{
Other timers: 
\code{\link{msg_tic}()}
}
\concept{timers}
