% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msg.R
\name{msg_tstamp}
\alias{msg_tstamp}
\title{StdErr timestamp with nice formatting}
\usage{
msg_tstamp(dt_format = "\%Y_\%m_\%d \%H:\%M:\%S \%Z", newlines = 2L)
}
\arguments{
\item{dt_format}{[chr] datetime format string : default "\%Y_\%m_\%d \%H:\%M:\%S \%Z"}

\item{newlines}{[int: default 2L] number of newlines to add before and
after message}
}
\value{
[none] side effect - displays to std_err
}
\description{
StdErr timestamp with nice formatting
}
\examples{
msg_tstamp()

}
