% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{order_draws}
\alias{order_draws}
\title{Order character draw names in order of their numeric indices.  PERD
compliant.}
\usage{
order_draws(draw_varnames, pe_varname = "point_estimate")
}
\arguments{
\item{draw_varnames}{[chr] vector of variable names}

\item{pe_varname}{[chr] name of point estimate variable, if present}
}
\value{
[chr] ordered vector of variable names
}
\description{
e.g. "draw_1", "draw_2", ..., "draw_10", "draw_11", ...
}
