% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_named_cli_args.R
\name{parse_all_named_cli_args}
\alias{parse_all_named_cli_args}
\title{Parse command-line arguments to global environment}
\usage{
parse_all_named_cli_args(
  trailingOnly = TRUE,
  assign_logical = TRUE,
  assign_integer = TRUE,
  assignment_env = globalenv()
)
}
\arguments{
\item{trailingOnly}{[lgl] if TRUE, only the trailing arguments are used -
passed to `commandArgs()`}

\item{assign_logical}{[lgl] if TRUE, assign true/false in parsed args to
`TRUE`/`FALSE`}

\item{assign_integer}{[lgl] if TRUE, assign solely numeric args to integer
e.g. 5 to `5L`}

\item{assignment_env}{[env] which environment to assign CLI arguments to as
named variables}
}
\value{
[named list] named list of CLI arguments
}
\description{
Add an arbitrary number of CLI args to a parser (default trailing args only),
then assign to a chosen environment (default `globalenv()`).  By default,
automatically assign logical and integer data types when they are detected
among CLI arguments, e.g. 'true' is assigned as `TRUE` and 5 is assigned as
`5L`.
}
\details{
This does not take advantage of argument defaults or types.  See
https://github.com/trevorld/r-argparse for documentation.
}
\examples{
\dontrun{
cli_args <- parse_all_named_cli_args(
   trailingOnly   = TRUE,
   assign_logical = TRUE,
   assign_integer = TRUE,
   assignment_env = globalenv()
)
}
}
