% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{PERD_regex}
\alias{PERD_regex}
\title{Regex to identify draw columns + point estimate + anything else the user deems helpful}
\usage{
PERD_regex(include_PE = TRUE, additions = NULL)
}
\value{
[chr] regex to identify draw columns + point estimate
}
\description{
Regex to identify draw columns + point estimate + anything else the user deems helpful
}
