% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_path.R
\name{prep_windows_path_for_unix}
\alias{prep_windows_path_for_unix}
\title{Convert windows file paths}
\usage{
prep_windows_path_for_unix(path = "clipboard")
}
\arguments{
\item{path}{[character] a filepath to convert}
}
\value{
correctly formatted filepath (`\` becomes `/`)
}
\description{
If you've copied a file path to the windows clipboard, call this function
without any arguments to both return a path with correct slashes, as well as
copy that path to the clipboard for an easy paste.
}
