% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msg.R
\name{prt_multiline}
\alias{prt_multiline}
\title{Paste data.frame output for a message}
\usage{
prt_multiline(x)
}
\arguments{
\item{x}{[list] list or data.frame to print}
}
\value{
[chr] data.frame output as a single string
}
\description{
Paste data.frame output for a message
}
\examples{
prt_multiline(head(mtcars, 2))
[1] "              mpg cyl disp  hp drat    wt  qsec vs am gear carb\nMazda RX4      21   6  160 110  3.9 2.620 16.46  0  1    4    4\nMazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4"
}
