% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubtab_source_meta.R
\name{pubtab_source_meta}
\alias{pubtab_source_meta}
\title{Generate a source metadata table}
\usage{
pubtab_source_meta(
  indicator,
  core_repo = "/share/code/geospatial/lbd_core/",
  resolve_urls = TRUE,
  out_file = NULL,
  return_output = TRUE
)
}
\arguments{
\item{indicator}{Name of the LBD indicator. This name should be related to
a prepped MBG dataset located in the following filepath:
`/share/geospatial/mbg/input_data/<indicator>.csv`.}

\item{core_repo}{[default='/share/code/geospatial/lbd_core/'] Path to
the core repository, used to load database reading packages.}

\item{resolve_urls}{[default=TRUE] Should GHDx links to each source be pulled?
(Specifying `TRUE` in this argument will slow down the function a bit)}

\item{out_file}{[default=NULL] If not `NULL`, saves the prepped source
metadata table to the specified filepath.}

\item{return_output}{[default=TRUE] If `TRUE`, returns the prepped source
metadata table as a data.table}
}
\value{
If `return_output==TRUE`, returns the prepped source metadata table
  as a data.table.
}
\description{
Generate an LBD source citation table for a publication SI.
}
\details{
Given a model-ready MBG dataset, generate a source citation table
  that contains critical metadata from the GHDx and sample size information
  gleaned from the input dataset. \strong{This function requires that the
  user has set up a valid ODBC file in their H:/ drive.} For more information,
  see \href{https://bit.ly/2x2AIZB}{the Hub documentation}.
}
\author{
Nat Henry, \email{nathenry@uw.edu}
}
