% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{read_file}
\alias{read_file}
\title{Read a file of an arbitrary type}
\usage{
read_file(
  path_to_file,
  verbose = FALSE,
  csv_opt = "readr::read_csv",
  return_DT = TRUE,
  ...
)
}
\arguments{
\item{path_to_file}{[chr] full path with extension}

\item{verbose}{[lgl] noisy or quiet function?}

\item{csv_opt}{[chr: default readr::read_csv] name spaced function call for csv reads}

\item{return_DT}{[lgl: default TRUE] return as.data.table?}

\item{...}{[any] additional arguments to pass to the reader function}
}
\value{
[file] an object of appropriate file type
}
\description{
Read a file of an arbitrary type
}
