% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet_partitioned.R
\name{read_parquet_partitioned}
\alias{read_parquet_partitioned}
\title{Read a partitioned parquet file structured as a directory with manifest file}
\usage{
read_parquet_partitioned(
  dir,
  fnames_ignored = c("manifest", "partition"),
  verbose = TRUE
)
}
\arguments{
\item{dir}{[chr] Directory containing partitioned parquet files e.g.
/path/to/stgpr/outputs/mv_20250101_words_words_words}

\item{fnames_ignored}{[chr: default c("manifest", "partition")] Vector of file names to ignore (e.g.,
manifest/partition files).}

\item{verbose}{[lgl: default TRUE] Print progress messages?}
}
\value{
[data.table] Combined data from all parquet files in the directory.
}
\description{
stgpr's api has changed, and final raked draws now stored as parquet instead
of csv
}
