% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dice.R
\name{roll}
\alias{roll}
\title{Roll dice and return a sampled value}
\usage{
roll(
  n_dice,
  n_sides,
  best_of_n = 1,
  method = c("max", "min"),
  plot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{n_dice}{e.g. 3 of 3d6}

\item{n_sides}{e.g 6 of 3d6}

\item{best_of_n}{if rolling 3d6 n times, choose the highest/lowest sum from n rolls}

\item{method}{"max" (advantage) or "min" (disadvantage)}

\item{plot}{if TRUE, display the distribution via dice()}
}
\value{
a single integer roll result
}
\description{
Roll dice and return a sampled value
}
\examples{
roll(2, 6) # just dice
roll(1, 20, best_of_n = 2, method = "max")  # advantage
roll(1, 20, best_of_n = 2, method = "min")  # disadvantage

stats <- list(
  str = integer()
, dex = integer()
, con = integer()
, int = integer()
, wis = integer()
, chr = integer()
)

# Roll stats with best of 2 & show plot
sapply(stats, function(x) roll(3,6,2,method='max',T))
# stat blocks for multiple NPCs/monsters
t(sapply(1:10, function(x) sapply(stats, function(x) roll(3,6)))) # normal monsters
t(sapply(1:10, function(x) sapply(stats, function(x) roll(3,6,2)))) # hard monsters
t(sapply(1:10, function(x) sapply(stats, function(x) roll(3,6,2,'min')))) # easy monsters
}
