% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_file}
\alias{save_file}
\title{SDG core function - save a file of approved format, with overwrite toggle}
\usage{
save_file(object, path_to_file, forbid_overwrite = TRUE, inform_user = FALSE)
}
\arguments{
\item{object}{[obj] any R object}

\item{path_to_file}{[path] duh}

\item{forbid_overwrite}{[lgl] default: overwrite not allowed}

\item{inform_user}{[lgl]}
}
\value{
[none] Saves to disk
}
\description{
SDG core function - save a file of approved format, with overwrite toggle
}
