% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_lancet.R
\name{set_magnitude}
\alias{set_magnitude}
\title{Define magnitude, magnitude label and denominator for a vector of numeric
values}
\usage{
set_magnitude(x, mag = NULL, allow_thousands = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{[num] numeric vector}

\item{mag}{[chr: default NULL {NULL, "B", "M"}] NULL (auto-detect), otherwise
user-override (not recommended) - (M)illions or (B)illions or (T)housands
(thousands are not Lancet-valid)}

\item{allow_thousands}{[lgl: default FALSE] allow (T)housands magnitude?  Not
Lancet-valid.}

\item{verbose}{[lgl: default TRUE] warn if allow_thousands is TRUE}
}
\value{
[list] with vector elements: mag, mag_label, and denom Each vector
  element has one item per length(x)
}
\description{
Support function used on _mean_ values for later formatting functions
}
\details{
Does not use 'thousands' label - not a valid Lancet format, which uses
ddd<narrow-space>ddd format.  See fround_mag_clu for that formatting
}
\examples{
set_magnitude(c(1, 1e3, 1e6))
# list(mag = c("", "", "M"), mag_label = c(" ", " ", " million "), denom = c(1, 1, 1e+06))
}
