% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journals.R
\name{thin_space}
\alias{thin_space}
\title{Thin space}
\usage{
thin_space()
}
\value{
[chr] thin space character (" ")
}
\description{
Lancet thin space standard (" ") for thousands separator (instead of commaa
",")
}
\examples{
thin_space()
}
