% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_of_parents.R
\name{validate_children_of_parents_inputs}
\alias{validate_children_of_parents_inputs}
\title{Helper function for children_of_parents.}
\usage{
validate_children_of_parents_inputs(parent_loc_ids, output, hierarchy)
}
\arguments{
\item{parent_loc_ids}{[int] ihme location ids}

\item{output}{[character] output options}

\item{hierarchy}{[data.table] ihme location hierarchy}
}
\value{
[none] stop on failure
}
\description{
Helper function for children_of_parents.
}
