% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents_of_children.R
\name{validate_parents_of_children_inputs}
\alias{validate_parents_of_children_inputs}
\title{Helper function to validate inputs to function}
\usage{
validate_parents_of_children_inputs(child_location_id, hierarchy, parent_level)
}
\arguments{
\item{child_location_id}{[int] ihme location_id}

\item{hierarchy}{[data.table] ihme location hierarchy}

\item{parent_level}{[int] ihme location level}
}
\value{
[none] stop on failure
}
\description{
Helper function to validate inputs to function
}
