% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{vec_to_comma_string}
\alias{vec_to_comma_string}
\title{Convert a vector to a comma-separated string}
\usage{
vec_to_comma_string(vec)
}
\arguments{
\item{vec}{[vector] some R vector}
}
\value{
[chr] a comma-separated string
}
\description{
Useful for passing R vectors to a CLI arg-parser
}
\examples{
vec_to_comma_string(1:10) # '1,2,3,4,5,6,7,8,9,10'
}
