% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_on_jobs.R
\name{wait_on_jobs}
\alias{wait_on_jobs}
\title{Given string pattern matching job name, wait while jobs are running or pending (SLURM)}
\usage{
wait_on_jobs(
  job_pattern,
  jobname_nchar = 50L,
  initial_sleep_sec = 5L,
  file_list = NULL,
  obj = NULL,
  resub = 0,
  perl = FALSE,
  dryrun = FALSE
)
}
\arguments{
\item{job_pattern}{[regex] a (perl) regular expression - submitted to `gnu grep` system commend}

\item{jobname_nchar}{[int] (defualt: 50) how many characters of a JobName to display and search}

\item{initial_sleep_sec}{[int] time to allow system time to find a submitted job}

\item{file_list}{[path/list] list of paths to files to wait on}

\item{obj}{[chr] checks if rhdf5::h5ls(file_list)$name exists}

\item{resub}{[int] if 0, return message about failures if missing_list has any items, else return 1}

\item{dryrun}{(default FALSE) if TRUE, only message and return system command, but do not submit}
}
\value{
[std_out] message to user
}
\description{
Given string pattern matching job name, wait while jobs are running or pending (SLURM)
}
\examples{
\dontrun{
wait_on_jobs(job_pattern = "my_job_name")
}
}
