% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_on_jobs.R
\name{wait_on_slurm_job_id}
\alias{wait_on_slurm_job_id}
\title{Wait for a Slurm job to finish}
\usage{
wait_on_slurm_job_id(job_id, initial_sleep = 5, perl = FALSE)
}
\arguments{
\item{job_id}{[int]}

\item{initial_sleep}{[int] default 5}

\item{perl}{[lgl] does you need perl to find your job_id?}
}
\value{
[std_out] console print
}
\description{
Wait for a Slurm job to finish
}
