% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_square.R
\name{assert_square}
\alias{assert_square}
\title{Assert a data.table is square by id_varnames}
\usage{
assert_square(
  dt,
  id_varnames,
  no_na_varnames = NULL,
  verbose = TRUE,
  hard_stop = TRUE
)
}
\arguments{
\item{dt}{[data.table] table of data}

\item{id_varnames}{[chr] vector of id variable names that, in combination (e.g. expand.grid) will uniquely ID all rows of data}

\item{no_na_varnames}{[chr] optional (defualt NULL) vector of variable names that must not be NA to be considered 'square'}

\item{verbose}{[lgl] print success message?}

\item{hard_stop}{[lgl] default TRUE - stop if non-square, warn if FALSE}
}
\value{
[list] 2 data.tables - duplicated rows and missing rows
}
\description{
Checks:
- duplicated rows by id_var combination
- missing rows by id_var combination
}
