% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents_of_children.R
\name{attach_national_location_id}
\alias{attach_national_location_id}
\title{Wrapper for attach_parent_location_id to attach national location_id (level 3)}
\usage{
attach_national_location_id(dt, hierarchy)
}
\arguments{
\item{dt}{[data.table] some table with columns `location_id`}

\item{hierarchy}{[data.table] ihme location hierarchy (get_location_metadata)}
}
\value{
[data.table] with new 'parent_location_id' column
}
\description{
Wrapper for attach_parent_location_id to attach national location_id (level 3)
}
