% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{build_metadata_shell}
\alias{build_metadata_shell}
\title{Build metadata from a template}
\usage{
build_metadata_shell(
  code_root = here::here(),
  jobname_filter = "^rst_ide|^vscode",
  submitline_n_char = 1000L,
  regex_to_extract = "singularity-images/rstudio/[[:graph:]]+\\\\.img",
  regex_to_ignore = "jpy",
  system_user_name = Sys.info()[["user"]],
  cluster_type = "slurm",
  verbose = FALSE
)
}
\arguments{
\item{code_root}{[path: default here::here()] Path to top-level code repo folder, require: path
contains a `.git` subfolder}

\item{jobname_filter}{[regex: default "^rst_ide|^vscode"] When you run `squeue -u <username>`, what
`JobName` indicates your interactive Rstudio session? (case-sensitive)
- see defaults}

\item{submitline_n_char}{[int: default 1000L] Length of submitted command string to expect
from system `squeue -j <jobid> -o submitline\%xxx` call (set this much
longer than you'd think necessary)}

\item{regex_to_extract}{[regex: default "singularity-images/rstudio/[[:graph:]]+\\.img"] What string do you want to extract after
running  `squeue -j <jobid> -o submitline\%xxx` (default - find Rstudio
image for provenance)}

\item{regex_to_ignore}{[regex: default "jpy"] If your `regex_to_extract` command finds more
strings than you want, this removes all strings matching this pattern}

\item{system_user_name}{[chr: default Sys.info()[["user"]]] User's identifier, according to the cluster}

\item{cluster_type}{[chr: default "slurm"] Only 'slurm' currently available - methods/calls
may differ by system in the future. (c ase-insensitive)}

\item{verbose}{[lgl: default FALSE] Do you want a std_err message of arguments to this
function?}
}
\value{
[list] Full metadata shell, including git info, cluster submission
  commands
}
\description{
Defaults currently designed for interactive Rstudio sessions.
}
\examples{
\dontrun{
metadata_shell <- build_metadata_shell()
}
}
