% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{build_metadata_shell}
\alias{build_metadata_shell}
\title{Build metadata from a template}
\usage{
build_metadata_shell(code_root, ...)
}
\arguments{
\item{code_root}{[path] path to top-level code repo folder, require: path
contains a `.git` subfolder}

\item{...}{[named vectors] any number of named vectors to append to the
metadata - be careful with this one}
}
\value{
[list] full metadata, including git info, cluster submission
  commands, and user-appended items
}
\description{
Build metadata from a template
}
