% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_of_parents.R
\name{children_of_parents}
\alias{children_of_parents}
\title{Get child locations given parent IDs from IHME hierarchies}
\usage{
children_of_parents(
  parent_loc_ids,
  hierarchy,
  output = "boolean",
  include_parent = FALSE
)
}
\arguments{
\item{parent_loc_ids}{Numeric vector of parent location_ids}

\item{hierarchy}{Which hierarchy to use? (Uses 'path_to_top_parent' to find all nested children)}

\item{output}{Output options:
\itemize{
 \item{\code{"boolean"} : T/F vector for all rows in the hierarchy - useful for adding a new column}
 \item{\code{"numeric"} : vector of only the childrens' location_ids (and parent, if desired)}
}}

\item{include_parent}{Logical T/F Do you want parents included with children?
(FALSE by default). If FALSE, and using two parent IDs, be careful if the
parent location_ids are nested inside each other.}
}
\value{
Logical T/F vector for all hierarchy locs, or a vector of children location_ids?
}
\description{
\code{children_of_parents()} finds children from IHME Hierarchies, given a
vector of location_ids.v From a hierarchy, which children are have a given
parent ID in the path_to_top_parent column?  Returns logical T/F vector.
}
\examples{

source(file.path("/ihme/cc_resources/libraries/current/r/get_location_metadata.R"))
hierarchy <- get_location_metadata(location_set_id = 111, location_set_version_id = 1050, release_id = 9)

hierarchy$children_of_china <- children_of_parents(6, hierarchy)
hierarchy$china_and_children <- children_of_parents(6, hierarchy, include_parent = TRUE)

child_locs_of_china <- children_of_parents(6, hierarchy, output = "loc_ids")

}
