% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msg.R
\name{datetime_stamp}
\alias{datetime_stamp}
\title{Current Date-Time Stamp}
\usage{
datetime_stamp(
  std_out = FALSE,
  std_err = FALSE,
  dt_format = "\%Y_\%m_\%d_\%H\%M\%S\%Z"
)
}
\arguments{
\item{std_out}{[std_out: default FALSE] using `print()`}

\item{std_err}{[std_err: default TRUE] message class}

\item{dt_format}{[chr] datetime format string default "\%Y_\%m_\%d_\%H\%M\%S\%Z"}
}
\value{
[invisible] date-time stamp
}
\description{
Wrapper for `Sys.time()`.  Results are formatted as YYYY_MM_DD_hhmmssTZONE.
Allows user to print to stderr and/or stdout, and invisibly returns the stamp.
}
\examples{
datetime_stamp(std_out = TRUE)

}
