% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{datetime_stamp}
\alias{datetime_stamp}
\title{Current Date-Time Stamp}
\usage{
datetime_stamp(
  std_out = TRUE,
  std_err = FALSE,
  dt_format = "\%Y_\%m_\%d_\%H\%M\%S\%Z"
)
}
\arguments{
\item{std_out}{[sdt_out] using `print()`}

\item{std_err}{[std_err] message class}
}
\value{
[invisible] date-time stamp
}
\description{
Wrapper for `Sys.time()`.  Results are formatted as YYYY_MM_DD_hhmmssTZONE.
Allows user to print to stderr, stdout, and invisible returns the stamp.
}
