% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_long_to_wide}
\alias{draws_long_to_wide}
\title{Transform draws from long to wide format by "draw_id"}
\usage{
draws_long_to_wide(
  DT,
  value_varname = "value",
  id_varnames = find_id_varnames(DT, removals = c(value_varname), verbose = FALSE),
  verbose = FALSE
)
}
\arguments{
\item{DT}{[data.table] input draws with columns `draw_id` and value_varname}

\item{value_varname}{[chr: default "value"] name of value variable in DT}

\item{id_varnames}{[character] columns to keep as is}

\item{verbose}{[lgl] print debug messages?}
}
\value{
[data.frame] draws in wide format
}
\description{
Draw columns will be ordered as c("point_estimate", "draw_0", "draw_1", "draw_2", ..., "draw_n")
}
