% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_long_to_wide}
\alias{draws_long_to_wide}
\title{Transform draws from long to wide format}
\usage{
draws_long_to_wide(
  DT,
  id_varnames = find_id_varnames(DT, removals = c("value"), verbose = FALSE),
  verbose = FALSE
)
}
\arguments{
\item{DT}{[data.table] input draws}

\item{id_varnames}{[character] columns to keep as is}

\item{verbose}{[lgl] print debug messages?}
}
\value{
[data.frame] draws in wide format
}
\description{
Draw columns will be ordered as c("point_estimate", "draw_0", "draw_1", "draw_2", ..., "draw_n")
}
