% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_to_mean_ci}
\alias{draws_to_mean_ci}
\title{Transform wide draws to mean and 95 percent CI}
\usage{
draws_to_mean_ci(
  DT,
  id_varnames = find_id_varnames(DT, verbose = FALSE),
  vars_draws_pe = find_draws_varnames(DT),
  remove_vars_draws = TRUE,
  remove_point_estimate = FALSE,
  remove_mean = FALSE,
  fix_mean_zero = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{DT}{[data.table] input draws in wide format}

\item{id_varnames}{[character] columns to keep as is}

\item{vars_draws_pe}{[chr] vector of variable names in `DT` that are draws + point_estimate}

\item{remove_vars_draws}{[lgl] remove draw columns?}

\item{remove_point_estimate}{[lgl] remove point_estimate column?}

\item{remove_mean}{[lgl] remove mean column?}

\item{fix_mean_zero}{[lgl] Some sets of draws have only a single value,
leading to a non-zero mean, and zeros in the UI.  This will set the mean to
zero if mean is > 0 and the upper is 0, or if mean < 0 and lower is 0.}

\item{verbose}{[lgl] print debug messages?}
}
\value{
[data.table] mean and 95 percent CI of draws (columns: mean, lower, upper),
  with or without draw columns, depending on `remove_vars_draws`
}
\description{
For now, retaining point estimate and mean by default since much processing
code depends on mean column
}
