% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{draws_years_to_wide}
\alias{draws_years_to_wide}
\title{Pivot long draws wide by years}
\usage{
draws_years_to_wide(DT, yr_vec = NULL)
}
\arguments{
\item{DT}{[data.table] input draws}
}
\value{
[data.table] draws in wide format
}
\description{
Pivot long draws wide by years
}
