% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{extract_cores}
\alias{extract_cores}
\title{Extract number of interactive user cores}
\usage{
extract_cores(
  system_user_name = user_name,
  jobname_filter = jobname_filter,
  cluster_type = cluster_type
)
}
\arguments{
\item{system_user_name}{[chr] how user is identified on the cluster}

\item{jobname_filter}{[regex] filter to include for `job_finder()` call}
}
\value{
[int] number of available cores for multithreading
- if user has more than one interactive session, defaults to 1
}
\description{
NOTE: This step finds cores with squeue, rather than sacct
}
