% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata_shell.R
\name{extract_submission_commands}
\alias{extract_submission_commands}
\title{Find a string in a user's recent Slurm squeue}
\usage{
extract_submission_commands(
  squeue_jobname_filter = "^rst_ide",
  max_cmd_length = 500L,
  string_to_extract = "ihme/singularity-images/rstudio/[:graph:]+",
  strings_to_ignore = "jpy",
  user_name = Sys.info()[["user"]]
)
}
\arguments{
\item{squeue_jobname_filter}{[character|regex] when you run `squeue -u
<username>`, what `NAME` do you want to filter for?}

\item{max_cmd_length}{[integer] how many characters long is your command?
Increase your default if the command is truncated.  All leading/trailing
whitespace is trimmed.}

\item{string_to_extract}{[character|regex] what string do you want to extract
after running  `sacct -j <jobid> -o submitline%xxx` using
`stringr::str_extract_all`}

\item{strings_to_ignore}{[character|regex] if your `string_to_extract`
command finds more strings than you want, this removes all strings with
this pattern anywhere inside using `stringr::str_detect`}

\item{user_name}{[character] which user's commands to find - defaults to your
own}
}
\value{
[list] all desired submission commands, and specific extracted text
  from string_to_extract
}
\description{
Intended to find Rstudio singularity image versions for metadata.  Finds a
user-defined string from the submission command you used to start your
Rstudio singularity image (or something else you desire).  Extracts this
informtion from ALL jobs you currently have active in your squeue.
}
