% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_transforms.R
\name{find_draws_varnames}
\alias{find_draws_varnames}
\title{Find draw variable names in a data.table}
\usage{
find_draws_varnames(DT, include_PE = TRUE, additions = NULL)
}
\arguments{
\item{DT}{[data.table]}

\item{include_PE}{[lgl: default TRUE] include point_estimate in search? Passed to PERD_regex()}

\item{additions}{[chr: default NULL] additional regex to include. Passed to PERD_regex()}
}
\value{
[chr] vector of variable names
}
\description{
Stops if no draw/PE columns found
}
