% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_lancet.R
\name{format_lancet_dt}
\alias{format_lancet_dt}
\title{Return a table with formatted central/lower/upper}
\usage{
format_lancet_dt(
  dt,
  d_type,
  central_var = "mean",
  lower_var = "lower",
  upper_var = "upper",
  remove_clu = TRUE,
  assert_clu_relationships = TRUE
)
}
\arguments{
\item{dt}{[data.table] with central/lower/upper columns}

\item{d_type}{[chr {prop', 'pp', 'count'}] data type - proportion, percentage point or count}

\item{central_var}{[chr: default 'mean'] name of central tendency e.g. 'point_estimate'}

\item{lower_var}{[chr: default 'lower']}

\item{upper_var}{[chr: default 'upper']}

\item{remove_clu}{[lgl: default TRUE] remove central/lower/upper columns?}

\item{assert_clu_relationships}{[lgl: default TRUE] enforce correct relationship between central/upper/lower?}
}
\value{
[data.table] with mean_95_UI_formatted column, and central/lower/upper columns removed (if specified)
}
\description{
Assumes a single data-type (d_type) for the whole table (e.g. 'prop', 'pp', 'count')
}
\examples{
DT <- data.table::data.table(location_did = 1, location_name = "Global", me_name = "vacc_dpt1",
                 mean = 55.8e6, lower = 50.7e6, upper = 60.7e6)
format_lancet_dt(dt = DT, d_type = "count", central_var = 'mean')
# location_did location_name   me_name     mean_95_UI_formatted
# <num>        <char>    <char>                   <char>
#   1:            1        Global vacc_dpt1 55·8 million (50·7–60·7)
}
