% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_lancet.R
\name{fround_dtype}
\alias{fround_dtype}
\title{Format and round with data-type suffix}
\usage{
fround_dtype(
  x,
  d_type = "prop",
  digits = 1L,
  nsmall = 1L,
  decimal.mark = mid_dot()
)
}
\arguments{
\item{x}{[num] numeric value}

\item{d_type}{[chr c('prop', 'pp', or 'count')] data type - proportion, percentage point or count}

\item{digits}{[integer] passed to `round()`}

\item{nsmall}{[integer] passed to `format()`}

\item{decimal.mark}{[chr (default mid_d)] decimal mark passed to `format()`}
}
\value{
[chr] formatted string
}
\description{
Format and round with data-type suffix
}
\examples{
fround_dtype(0.123456789) # "0·1\%"
}
