% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{get_new_output_dv}
\alias{get_new_output_dv}
\title{Increment a new output folder date-version}
\usage{
get_new_output_dv(root, date = "today")
}
\arguments{
\item{root}{[chr] path to root of output results}

\item{date}{[chr] character date in form of "YYYY_MM_DD" or "today". "today" will be interpreted as today's date.}
}
\value{
[dbl] new output version
}
\description{
Increment a new output folder date-version
}
\examples{
get_new_output_dv(root = tempdir(), date = "today")
}
