% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_x.R
\name{is_empty}
\alias{is_empty}
\title{Check that a vector/element is not empty}
\usage{
is_empty(x, check_whitespace = TRUE)
}
\arguments{
\item{x}{[vec] some scalar element}
}
\value{
[lgl]
}
\description{
Checks for length == 1, is.null, is.na, empty string ("")
}
