% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_output_dir.R
\name{make_new_output_dir}
\alias{make_new_output_dir}
\title{Make a date-versioned output directory for output results.}
\usage{
make_new_output_dir(root, date = "today", suffix = "")
}
\arguments{
\item{root}{[chr] path to root of output results}

\item{date}{[chr] character date in form of "YYYY_MM_DD" or "today". "today" will be interpreted as today's date.}

\item{suffix}{[chr] a suffix to append after the incremented version number index}
}
\value{
[chr] path to new output directory of the form "/<root>/YYYY_MM_DD.VV"
}
\description{
Returns an appropriate path to save results in, creating it if necessary.
}
\examples{
\dontrun{
make_new_output_dir(tempdir(), date = "today")
}
}
