% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_msg.R
\name{msg_tic}
\alias{msg_tic}
\title{Start tictoc timer}
\usage{
msg_tic()
}
\value{
[none]
}
\description{
Just a wrapper of tictoc::tic() for fun name consistency
}
\examples{
msg_tic()
msg_toc()
}
\seealso{
Other timers: 
\code{\link{msg_toc}()}
}
\concept{timers}
