% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{read_file}
\alias{read_file}
\title{Read a file of an arbitrary type}
\usage{
read_file(path_to_file, verbose = FALSE, csv_opt = "data.table::fread", ...)
}
\arguments{
\item{path_to_file}{[chr] full path with extenstion}

\item{verbose}{[lgl] noisy or quiet function?}

\item{csv_opt}{[chr] namespaced function call for csv reads (default `"data.table::fread"`)}

\item{...}{[any] additional arguments to pass to the reader function}
}
\value{
[file] an object of appropriate file type
}
\description{
Read a file of an arbitrary type
}
