% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{read_file}
\alias{read_file}
\title{Read a file of approved format}
\usage{
read_file(f_path, verbose = FALSE)
}
\arguments{
\item{f_path}{[path]}

\item{verbose}{[lgl]}
}
\value{
[obj] use this to assign a file to an R object
}
\description{
Read a file of approved format
}
