% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_file}
\alias{save_file}
\title{Save file I/O with overwrite and message control}
\usage{
save_file(object, f_path, forbid_overwrite = TRUE, verbose = FALSE)
}
\arguments{
\item{object}{[obj] any R object}

\item{f_path}{[path] duh}

\item{forbid_overwrite}{[lgl] default: overwrite not allowed}

\item{verbose}{[lgl] default: silent}
}
\value{
[none] Saves to disk
}
\description{
Save file I/O with overwrite and message control
}
