% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_file}
\alias{save_file}
\title{Save file I/O with overwrite and message control}
\usage{
save_file(
  object,
  f_path,
  overwrite = NULL,
  verbose = FALSE,
  forbid_overwrite = NULL
)
}
\arguments{
\item{object}{[obj] an R object}

\item{f_path}{[path] a file path to save to}

\item{overwrite}{[lgl] default: NULL which overwrites by default until forbid_overwrite is deprecated}

\item{verbose}{[lgl] default: silent}

\item{forbid_overwrite}{[lgl]  DEPRECATED - backward compatibility preserved}
}
\value{
[none] Saves to disk
}
\description{
Defualt behavior is to overwrite
}
